+++ .github/actions/check-for-CLA/index.js
@@ -32,7 +32,7 @@ const main = async () => {
 
   const response = await postCommentOnPullRequest(
     hasSignedCLA,
-    errorFoundOnCLACheck
+    errorFoundOnCLACheck,
   );
 };
 
@@ -49,7 +49,7 @@ const checkIfUserHasSignedAnyCLA = async () => {
 const checkIfIndividualCLAFound = async () => {
   const response = await getValuesFromGoogleSheet(
     GOOGLE_SHEETS_INFO.individualCLASheetId,
-    "D2:D"
+    "D2:D",
   );
 
   const rows = response.data.values;
@@ -70,7 +70,7 @@ const checkIfIndividualCLAFound = async () => {
 const checkIfCorporateCLAFound = async () => {
   const response = await getValuesFromGoogleSheet(
     GOOGLE_SHEETS_INFO.corporateCLASheetId,
-    "H2:H"
+    "H2:H",
   );
 
   const rows = response.data.values;
@@ -132,14 +132,14 @@ const postCommentOnPullRequest = async (hasSignedCLA, errorFoundOnCLACheck) => {
         accept: "application/vnd.github+json",
         "X-GitHub-Api-Version": "2022-11-28",
       },
-    }
+    },
   );
 };
 
 const getCommentBody = (hasSignedCLA, errorFoundOnCLACheck) => {
   const commentTemplate = fs.readFileSync(
     "./.github/actions/check-for-CLA/templates/pullRequestComment.hbs",
-    "utf-8"
+    "utf-8",
   );
 
   const getCommentFromTemplate = Handlebars.compile(commentTemplate);