+++ packages/engine/Source/Scene/Scene.js
@@ -72,6 +72,7 @@ import SunLight from "./SunLight.js";
 import SunPostProcess from "./SunPostProcess.js";
 import TweenCollection from "./TweenCollection.js";
 import View from "./View.js";
+import VRPose from "./VRPose.js";
 import DebugInspector from "./DebugInspector.js";
 import VoxelCell from "./VoxelCell.js";
 import VoxelPrimitive from "./VoxelPrimitive.js";
@@ -154,7 +155,7 @@ function Scene(options) {
   this._frameState = new FrameState(
     context,
     new CreditDisplay(creditContainer, " • ", creditViewport),
-    this._jobScheduler
+    this._jobScheduler,
   );
   this._frameState.scene3DOnly = defaultValue(options.scene3DOnly, false);
   this._removeCreditContainer = !hasCreditContainer;
@@ -564,7 +565,7 @@ function Scene(options) {
   this.invertClassificationColor = Color.clone(Color.WHITE);
 
   this._actualInvertClassificationColor = Color.clone(
-    this._invertClassificationColor
+    this._invertClassificationColor,
   );
   this._invertClassification = new InvertClassification();
 
@@ -622,11 +623,13 @@ function Scene(options) {
     usePostProcess: false,
     usePostProcessSelected: false,
     useWebVR: false,
+    useWebXR: true,
   };
 
   this._useWebVR = false;
-  this._cameraVR = undefined;
-  this._aspectRatioVR = undefined;
+  this._useWebXR = true;
+  this._webXR = undefined;
+  this._poseVR = undefined;
 
   /**
    * When <code>true</code>, rendering a frame will only occur when needed as determined by changes within the scene.
@@ -660,17 +663,19 @@ function Scene(options) {
    */
   this.maximumRenderTimeChange = defaultValue(
     options.maximumRenderTimeChange,
-    0.0
+    0.0,
   );
   this._lastRenderTime = undefined;
   this._frameRateMonitor = undefined;
 
-  this._removeRequestListenerCallback = RequestScheduler.requestCompletedEvent.addEventListener(
-    requestRenderAfterFrame(this)
-  );
-  this._removeTaskProcessorListenerCallback = TaskProcessor.taskCompletedEvent.addEventListener(
-    requestRenderAfterFrame(this)
-  );
+  this._removeRequestListenerCallback =
+    RequestScheduler.requestCompletedEvent.addEventListener(
+      requestRenderAfterFrame(this),
+    );
+  this._removeTaskProcessorListenerCallback =
+    TaskProcessor.taskCompletedEvent.addEventListener(
+      requestRenderAfterFrame(this),
+    );
   this._removeGlobeCallbacks = [];
   this._removeTerrainProviderReadyListener = undefined;
 
@@ -678,7 +683,7 @@ function Scene(options) {
     0,
     0,
     context.drawingBufferWidth,
-    context.drawingBufferHeight
+    context.drawingBufferHeight,
   );
   const camera = new Camera(this);
 
@@ -751,13 +756,13 @@ function updateGlobeListeners(scene, globe) {
   if (defined(globe)) {
     removeGlobeCallbacks.push(
       globe.imageryLayersUpdatedEvent.addEventListener(
-        requestRenderAfterFrame(scene)
-      )
+        requestRenderAfterFrame(scene),
+      ),
     );
     removeGlobeCallbacks.push(
       globe.terrainProviderChanged.addEventListener(
-        requestRenderAfterFrame(scene)
-      )
+        requestRenderAfterFrame(scene),
+      ),
     );
   }
   scene._removeGlobeCallbacks = removeGlobeCallbacks;
@@ -1368,7 +1373,7 @@ Object.defineProperties(Scene.prototype, {
       //>>includeStart('debug', pragmas.debug);
       if (this.scene3DOnly && value !== SceneMode.SCENE3D) {
         throw new DeveloperError(
-          "Only SceneMode.SCENE3D is valid when scene3DOnly is true."
+          "Only SceneMode.SCENE3D is valid when scene3DOnly is true.",
         );
       }
       //>>includeEnd('debug');
@@ -1381,7 +1386,7 @@ Object.defineProperties(Scene.prototype, {
         //>>includeStart('debug', pragmas.debug);
       } else {
         throw new DeveloperError(
-          "value must be a valid SceneMode enumeration."
+          "value must be a valid SceneMode enumeration.",
         );
         //>>includeEnd('debug');
       }
@@ -1417,7 +1422,7 @@ Object.defineProperties(Scene.prototype, {
 
   /**
    * When <code>true</code>, splits the scene into two viewports with steroscopic views for the left and right eyes.
-   * Used for cardboard and WebVR.
+   * Used for cardboard, WebVR and WebXR.
    * @memberof Scene.prototype
    * @type {boolean}
    * @default false
@@ -1430,32 +1435,86 @@ Object.defineProperties(Scene.prototype, {
       //>>includeStart('debug', pragmas.debug);
       if (this.camera.frustum instanceof OrthographicFrustum) {
         throw new DeveloperError(
-          "VR is unsupported with an orthographic projection."
+          "VR is unsupported with an orthographic projection.",
         );
       }
       //>>includeEnd('debug');
       this._useWebVR = value;
       if (this._useWebVR) {
         this._frameState.creditDisplay.container.style.visibility = "hidden";
-        this._cameraVR = new Camera(this);
+        this._poseVR = new VRPose(this);
         if (!defined(this._deviceOrientationCameraController)) {
-          this._deviceOrientationCameraController = new DeviceOrientationCameraController(
-            this
-          );
+          this._deviceOrientationCameraController =
+            new DeviceOrientationCameraController(this);
         }
-
-        this._aspectRatioVR = this.camera.frustum.aspectRatio;
       } else {
         this._frameState.creditDisplay.container.style.visibility = "visible";
-        this._cameraVR = undefined;
+        this._poseVR = undefined;
         this._deviceOrientationCameraController =
           this._deviceOrientationCameraController &&
           !this._deviceOrientationCameraController.isDestroyed() &&
           this._deviceOrientationCameraController.destroy();
-
-        this.camera.frustum.aspectRatio = this._aspectRatioVR;
-        this.camera.frustum.xOffset = 0.0;
       }
+    },
+  },
+
+  /**
+   * When <code>false</code>, ignores the presence of the WebXR API and falls back to WebVR behavior.
+   * @memberof Scene.prototype
+   * @type {boolean}
+   * @default true
+   */
+  useWebXR: {
+    get: function () {
+      return this._useWebXR;
+    },
+    set: function (value) {
+      this._useWebXR = value;
+    },
+  },
+
+  /**
+   * An object from which WebXR information is pulled.
+   * <p>
+   * Member <code>refSpace</code> should be set when
+   * <code>XRSession:requestReferenceSpace</code> is resolved. If
+   * <code>refSpace</code> is not set, an inline reference space will
+   * be assumed.
+   * </p>
+   * <p>
+   * Member <code>frame</code> must be set with the
+   * <code>XRFrame</code> parameter received at the callback scheduled
+   * with <code>XRSession:requestAnimationFrame</code>.
+   * </p>
+   *
+   * @memberof Scene.prototype
+   * @type {object}
+   * @default undefined
+   *
+   * @see Scene#useWebVR
+   * @see {@link https://developer.mozilla.org/en-US/docs/Web/API/XRSession/requestReferenceSpace|requestReferenceSpace}
+   * @see {@link https://developer.mozilla.org/en-US/docs/Web/API/XRSession/requestAnimationFrame|requestAnimationFrame}
+   *
+   * @example
+   * const scene = new Cesium.Scene({ canvas: canvas });
+   * scene.useWebVR = true;
+   * scene.webXRContext = { refSpace: undefined, frame: undefined };
+   * navigator.xr.requestSession("immersive-vr").then((xrSession) => {
+   *   xrSession.requestReferenceSpace("local").then((xrRefSpace) => {
+   *     scene.webXRContext.refSpace = xrRefSpace;
+   *     xrSession.requestAnimationFrame((time, xrFrame) => {
+   *       scene.webXRContext.frame = xrFrame;
+   *       scene.render();
+   *     });
+   *   });
+   * });
+   */
+  webXRContext: {
+    get: function () {
+      return this._webXR;
+    },
+    set: function (value) {
+      this._webXR = value;
     },
   },
 
@@ -1503,7 +1562,7 @@ Object.defineProperties(Scene.prototype, {
       //>>includeStart('debug', pragmas.debug);
       if (!defined(value) || value < 0.0) {
         throw new DeveloperError(
-          "minimumDisableDepthTestDistance must be greater than or equal to 0.0."
+          "minimumDisableDepthTestDistance must be greater than or equal to 0.0.",
         );
       }
       //>>includeEnd('debug');
@@ -1698,7 +1757,7 @@ function updateDerivedCommands(scene, command, shadowsDirty) {
       scene,
       command,
       context,
-      derivedCommands.picking
+      derivedCommands.picking,
     );
   }
 
@@ -1707,7 +1766,7 @@ function updateDerivedCommands(scene, command, shadowsDirty) {
       scene,
       command,
       context,
-      derivedCommands.depth
+      derivedCommands.depth,
     );
   }
 
@@ -1717,7 +1776,7 @@ function updateDerivedCommands(scene, command, shadowsDirty) {
     derivedCommands.hdr = DerivedCommand.createHdrCommand(
       command,
       context,
-      derivedCommands.hdr
+      derivedCommands.hdr,
     );
     command = derivedCommands.hdr.command;
     derivedCommands = command.derivedCommands;
@@ -1729,7 +1788,7 @@ function updateDerivedCommands(scene, command, shadowsDirty) {
       command,
       shadowsDirty,
       context,
-      derivedCommands.shadows
+      derivedCommands.shadows,
     );
   }
 
@@ -1741,13 +1800,13 @@ function updateDerivedCommands(scene, command, shadowsDirty) {
       derivedCommands.oit.shadows = oit.createDerivedCommands(
         derivedCommands.shadows.receiveCommand,
         context,
-        derivedCommands.oit.shadows
+        derivedCommands.oit.shadows,
       );
     } else {
       derivedCommands.oit = oit.createDerivedCommands(
         command,
         context,
-        derivedCommands.oit
+        derivedCommands.oit,
       );
     }
   }
@@ -1801,7 +1860,7 @@ Scene.prototype.updateDerivedCommands = function (command) {
         command,
         shadowsDirty,
         context,
-        derivedCommands.shadows
+        derivedCommands.shadows,
       );
     }
 
@@ -1809,12 +1868,12 @@ Scene.prototype.updateDerivedCommands = function (command) {
       derivedCommands.logDepth = DerivedCommand.createLogDepthCommand(
         command,
         context,
-        derivedCommands.logDepth
+        derivedCommands.logDepth,
       );
       updateDerivedCommands(
         this,
         derivedCommands.logDepth.command,
-        shadowsDirty
+        shadowsDirty,
       );
     }
     if (hasDerivedCommands || needsDerivedCommands) {
@@ -1860,7 +1919,7 @@ function getOccluder(scene) {
     scratchOccluder = Occluder.fromBoundingSphere(
       scratchOccluderBoundingSphere,
       scene.camera.positionWC,
-      scratchOccluder
+      scratchOccluder,
     );
     return scratchOccluder;
   }
@@ -1905,11 +1964,12 @@ Scene.prototype.updateFrameState = function () {
   frameState.cullingVolume = camera.frustum.computeCullingVolume(
     camera.positionWC,
     camera.directionWC,
-    camera.upWC
+    camera.upWC,
   );
   frameState.occluder = getOccluder(this);
   frameState.minimumTerrainHeight = 0.0;
-  frameState.minimumDisableDepthTestDistance = this._minimumDisableDepthTestDistance;
+  frameState.minimumDisableDepthTestDistance =
+    this._minimumDisableDepthTestDistance;
   frameState.invertClassification = this.invertClassification;
   frameState.useLogDepth =
     this._logDepthBuffer &&
@@ -1931,14 +1991,17 @@ Scene.prototype.updateFrameState = function () {
     globe._terrainExaggerationChanged = false;
   }
   frameState.verticalExaggeration = this.verticalExaggeration;
-  frameState.verticalExaggerationRelativeHeight = this.verticalExaggerationRelativeHeight;
+  frameState.verticalExaggerationRelativeHeight =
+    this.verticalExaggerationRelativeHeight;
 
   if (
     defined(this._specularEnvironmentMapAtlas) &&
     this._specularEnvironmentMapAtlas.ready
   ) {
-    frameState.specularEnvironmentMaps = this._specularEnvironmentMapAtlas.texture;
-    frameState.specularEnvironmentMapsMaximumLOD = this._specularEnvironmentMapAtlas.maximumMipmapLevel;
+    frameState.specularEnvironmentMaps =
+      this._specularEnvironmentMapAtlas.texture;
+    frameState.specularEnvironmentMapsMaximumLOD =
+      this._specularEnvironmentMapAtlas.maximumMipmapLevel;
   } else {
     frameState.specularEnvironmentMaps = undefined;
     frameState.specularEnvironmentMapsMaximumLOD = undefined;
@@ -1948,7 +2011,7 @@ Scene.prototype.updateFrameState = function () {
 
   this._actualInvertClassificationColor = Color.clone(
     this.invertClassificationColor,
-    this._actualInvertClassificationColor
+    this._actualInvertClassificationColor,
   );
   if (!InvertClassification.isTranslucencySupported(this._context)) {
     this._actualInvertClassificationColor.alpha = 1.0;
@@ -1999,11 +2062,11 @@ let transformFrom2D = new Matrix4(
   0.0,
   0.0,
   0.0,
-  1.0
+  1.0,
 );
 transformFrom2D = Matrix4.inverseTransformation(
   transformFrom2D,
-  transformFrom2D
+  transformFrom2D,
 );
 
 function debugShowBoundingVolume(command, scene, passState, debugFramebuffer) {
@@ -2035,8 +2098,8 @@ function debugShowBoundingVolume(command, scene, passState, debugFramebuffer) {
         new EllipsoidGeometry({
           radii: new Cartesian3(radius, radius, radius),
           vertexFormat: PerInstanceColorAppearance.FLAT_VERTEX_FORMAT,
-        })
-      )
+        }),
+      ),
     );
 
     scene._debugVolume = new Primitive({
@@ -2061,8 +2124,8 @@ function debugShowBoundingVolume(command, scene, passState, debugFramebuffer) {
         BoxGeometry.fromDimensions({
           dimensions: new Cartesian3(2.0, 2.0, 2.0),
           vertexFormat: PerInstanceColorAppearance.FLAT_VERTEX_FORMAT,
-        })
-      )
+        }),
+      ),
     );
 
     scene._debugVolume = new Primitive({
@@ -2071,7 +2134,7 @@ function debugShowBoundingVolume(command, scene, passState, debugFramebuffer) {
         modelMatrix: Matrix4.fromRotationTranslation(
           halfAxes,
           center,
-          new Matrix4()
+          new Matrix4(),
         ),
         attributes: {
           color: new ColorGeometryInstanceAttribute(1.0, 0.0, 0.0, 1.0),
@@ -2163,7 +2226,7 @@ function executeCommand(command, scene, context, passState, debugFramebuffer) {
     scene._debugInspector.executeDebugShowFrustumsCommand(
       scene,
       command,
-      passState
+      passState,
     );
     return;
   }
@@ -2224,7 +2287,7 @@ function executeTranslucentCommandsBackToFront(
   executeFunction,
   passState,
   commands,
-  invertClassification
+  invertClassification,
 ) {
   const context = scene.context;
 
@@ -2235,7 +2298,7 @@ function executeTranslucentCommandsBackToFront(
       invertClassification.unclassifiedCommand,
       scene,
       context,
-      passState
+      passState,
     );
   }
 
@@ -2250,7 +2313,7 @@ function executeTranslucentCommandsFrontToBack(
   executeFunction,
   passState,
   commands,
-  invertClassification
+  invertClassification,
 ) {
   const context = scene.context;
 
@@ -2261,7 +2324,7 @@ function executeTranslucentCommandsFrontToBack(
       invertClassification.unclassifiedCommand,
       scene,
       context,
-      passState
+      passState,
     );
   }
 
@@ -2332,7 +2395,7 @@ function executeCommands(scene, passState) {
         environmentState.skyAtmosphereCommand,
         scene,
         context,
-        passState
+        passState,
       );
     }
 
@@ -2368,7 +2431,7 @@ function executeCommands(scene, passState) {
         executeFunction,
         passState,
         commands,
-        invertClassification
+        invertClassification,
       ) {
         view.globeDepth.prepareColorTextures(context);
         view.oit.executeCommands(
@@ -2376,7 +2439,7 @@ function executeCommands(scene, passState) {
           executeFunction,
           passState,
           commands,
-          invertClassification
+          invertClassification,
         );
       };
     }
@@ -2443,7 +2506,7 @@ function executeCommands(scene, passState) {
         executeCommand,
         globeTranslucencyFramebuffer,
         scene,
-        passState
+        passState,
       );
     } else {
       for (j = 0; j < length; ++j) {
@@ -2468,7 +2531,7 @@ function executeCommands(scene, passState) {
           executeCommand,
           globeTranslucencyFramebuffer,
           scene,
-          passState
+          passState,
         );
       } else {
         for (j = 0; j < length; ++j) {
@@ -2508,7 +2571,7 @@ function executeCommands(scene, passState) {
             context,
             passState,
             clearGlobeDepth,
-            globeDepth.depthStencilTexture
+            globeDepth.depthStencilTexture,
           );
         }
 
@@ -2575,7 +2638,7 @@ function executeCommands(scene, passState) {
           context,
           passState,
           clearGlobeDepth,
-          scene._invertClassification._fbo.getDepthStencilTexture()
+          scene._invertClassification._fbo.getDepthStencilTexture(),
         );
       }
 
@@ -2656,7 +2719,7 @@ function executeCommands(scene, passState) {
       executeCommand,
       passState,
       commands,
-      invertClassification
+      invertClassification,
     );
 
     // Classification for translucent 3D Tiles
@@ -2671,13 +2734,13 @@ function executeCommands(scene, passState) {
         executeCommand,
         passState,
         commands,
-        globeDepth.depthStencilTexture
+        globeDepth.depthStencilTexture,
       );
       view.translucentTileClassification.executeClassificationCommands(
         scene,
         executeCommand,
         passState,
-        frustumCommands
+        frustumCommands,
       );
     }
 
@@ -2719,7 +2782,7 @@ function executeCommands(scene, passState) {
         executeIdCommand,
         globeTranslucencyFramebuffer,
         scene,
-        passState
+        passState,
       );
     } else {
       for (j = 0; j < length; ++j) {
@@ -2878,21 +2941,19 @@ function executeShadowMapCastCommands(scene) {
           command.derivedCommands.shadows.castCommands[i],
           scene,
           context,
-          pass.passState
+          pass.passState,
         );
       }
     }
   }
 }
 
-const scratchEyeTranslation = new Cartesian3();
-
 /**
  * @private
  */
 Scene.prototype.updateAndExecuteCommands = function (
   passState,
-  backgroundColor
+  backgroundColor,
 ) {
   const frameState = this._frameState;
   const mode = frameState.mode;
@@ -2913,64 +2974,62 @@ Scene.prototype.updateAndExecuteCommands = function (
 
 function executeWebVRCommands(scene, passState, backgroundColor) {
   const view = scene._view;
-  const camera = view.camera;
   const environmentState = scene._environmentState;
-  const renderTranslucentDepthForPick =
-    environmentState.renderTranslucentDepthForPick;
+
+  const camera = view.camera;
+  const pose = scene._poseVR;
+  if (!pose.prepare(camera, passState.viewport)) {
+    // TODO: we should handle this gracefully.
+    // No need to do any further processing if a pose was not obtained.
+    return;
+  }
 
   updateAndClearFramebuffers(scene, passState, backgroundColor);
-
   updateAndRenderPrimitives(scene);
-
   view.createPotentiallyVisibleSet(scene);
-
   executeComputeCommands(scene);
-
-  if (!renderTranslucentDepthForPick) {
+  if (!environmentState.renderTranslucentDepthForPick) {
     executeShadowMapCastCommands(scene);
   }
 
-  // Based on Calculating Stereo pairs by Paul Bourke
-  // http://paulbourke.net/stereographics/stereorender/
-  const viewport = passState.viewport;
-  viewport.x = 0;
-  viewport.y = 0;
-  viewport.width = viewport.width * 0.5;
+  pose.apply(() => {
+    executeCommands(scene, passState);
+  });
 
-  const savedCamera = Camera.clone(camera, scene._cameraVR);
-  savedCamera.frustum = camera.frustum;
+  // Now manually blit into the framebuffer provided by WebXR, which
+  // is the framebuffer for the HMD's displays.
 
-  const near = camera.frustum.near;
-  const fo = near * defaultValue(scene.focalLength, 5.0);
-  const eyeSeparation = defaultValue(scene.eyeSeparation, fo / 30.0);
-  const eyeTranslation = Cartesian3.multiplyByScalar(
-    savedCamera.right,
-    eyeSeparation * 0.5,
-    scratchEyeTranslation
-  );
-
-  camera.frustum.aspectRatio = viewport.width / viewport.height;
-
-  const offset = (0.5 * eyeSeparation * near) / fo;
-
-  Cartesian3.add(savedCamera.position, eyeTranslation, camera.position);
-  camera.frustum.xOffset = offset;
-
-  executeCommands(scene, passState);
-
-  viewport.x = viewport.width;
-
-  Cartesian3.subtract(savedCamera.position, eyeTranslation, camera.position);
-  camera.frustum.xOffset = -offset;
-
-  executeCommands(scene, passState);
-
-  Camera.clone(savedCamera, camera);
+  // When the WebXR Emulator is used, xrLayer has no framebuffer.
+  if (
+    pose.isWebXR &&
+    defined(pose.xrLayer) &&
+    defined(pose.xrLayer.framebuffer)
+  ) {
+    const xrLayer = pose.xrLayer;
+    const srcFb = passState.framebuffer;
+    const gl = srcFb._gl;
+    gl.bindFramebuffer(gl.READ_FRAMEBUFFER, srcFb._framebuffer);
+    gl.bindFramebuffer(gl.DRAW_FRAMEBUFFER, xrLayer.framebuffer);
+    gl.blitFramebuffer(
+      0,
+      0,
+      xrLayer.framebufferWidth,
+      xrLayer.framebufferHeight,
+      0,
+      0,
+      xrLayer.framebufferWidth,
+      xrLayer.framebufferHeight,
+      gl.COLOR_BUFFER_BIT,
+      gl.NEAREST,
+    );
+    gl.bindFramebuffer(gl.READ_FRAMEBUFFER, null);
+    gl.bindFramebuffer(gl.DRAW_FRAMEBUFFER, null);
+  }
 }
 
 const scratch2DViewportCartographic = new Cartographic(
   Math.PI,
-  CesiumMath.PI_OVER_TWO
+  CesiumMath.PI_OVER_TWO,
 );
 const scratch2DViewportMaxCoord = new Cartesian3();
 const scratch2DViewportSavedPosition = new Cartesian3();
@@ -2997,11 +3056,11 @@ function execute2DViewportCommands(scene, passState) {
 
   const position = Cartesian3.clone(
     camera.position,
-    scratch2DViewportSavedPosition
+    scratch2DViewportSavedPosition,
   );
   const transform = Matrix4.clone(
     camera.transform,
-    scratch2DViewportCameraTransform
+    scratch2DViewportCameraTransform,
   );
   const frustum = camera.frustum.clone();
 
@@ -3011,7 +3070,7 @@ function execute2DViewportCommands(scene, passState) {
     viewport,
     0.0,
     1.0,
-    scratch2DViewportTransform
+    scratch2DViewportTransform,
   );
   const projectionMatrix = camera.frustum.projectionMatrix;
 
@@ -3020,13 +3079,13 @@ function execute2DViewportCommands(scene, passState) {
     CesiumMath.sign(x) * maxCoord.x - x,
     0.0,
     -camera.positionWC.x,
-    scratch2DViewportEyePoint
+    scratch2DViewportEyePoint,
   );
   const windowCoordinates = Transforms.pointToGLWindowCoordinates(
     projectionMatrix,
     viewportTransformation,
     eyePoint,
-    scratch2DViewportWindowCoords
+    scratch2DViewportWindowCoords,
   );
 
   windowCoordinates.x = Math.floor(windowCoordinates.x);
@@ -3052,7 +3111,7 @@ function execute2DViewportCommands(scene, passState) {
     frameState.cullingVolume = camera.frustum.computeCullingVolume(
       camera.positionWC,
       camera.directionWC,
-      camera.upWC
+      camera.upWC,
     );
     context.uniformState.update(frameState);
 
@@ -3068,7 +3127,7 @@ function execute2DViewportCommands(scene, passState) {
     frameState.cullingVolume = camera.frustum.computeCullingVolume(
       camera.positionWC,
       camera.directionWC,
-      camera.upWC
+      camera.upWC,
     );
     context.uniformState.update(frameState);
 
@@ -3082,7 +3141,7 @@ function execute2DViewportCommands(scene, passState) {
     frameState.cullingVolume = camera.frustum.computeCullingVolume(
       camera.positionWC,
       camera.directionWC,
-      camera.upWC
+      camera.upWC,
     );
     context.uniformState.update(frameState);
 
@@ -3099,7 +3158,7 @@ function execute2DViewportCommands(scene, passState) {
     frameState.cullingVolume = camera.frustum.computeCullingVolume(
       camera.positionWC,
       camera.directionWC,
-      camera.upWC
+      camera.upWC,
     );
     context.uniformState.update(frameState);
 
@@ -3114,7 +3173,7 @@ function execute2DViewportCommands(scene, passState) {
     frameState.cullingVolume = camera.frustum.computeCullingVolume(
       camera.positionWC,
       camera.directionWC,
-      camera.upWC
+      camera.upWC,
     );
     context.uniformState.update(frameState);
 
@@ -3131,7 +3190,7 @@ function execute2DViewportCommands(scene, passState) {
     frameState.cullingVolume = camera.frustum.computeCullingVolume(
       camera.positionWC,
       camera.directionWC,
-      camera.upWC
+      camera.upWC,
     );
     context.uniformState.update(frameState);
 
@@ -3148,7 +3207,7 @@ function executeCommandsInViewport(
   firstViewport,
   scene,
   passState,
-  backgroundColor
+  backgroundColor,
 ) {
   const environmentState = scene._environmentState;
   const view = scene._view;
@@ -3209,7 +3268,7 @@ Scene.prototype.updateEnvironment = function () {
     if (defined(skyAtmosphere)) {
       if (defined(globe)) {
         skyAtmosphere.setDynamicLighting(
-          DynamicAtmosphereLightingType.fromGlobeFlags(globe)
+          DynamicAtmosphereLightingType.fromGlobeFlags(globe),
         );
         environmentState.isReadyForAtmosphere =
           environmentState.isReadyForAtmosphere ||
@@ -3223,7 +3282,7 @@ Scene.prototype.updateEnvironment = function () {
 
       environmentState.skyAtmosphereCommand = skyAtmosphere.update(
         frameState,
-        globe
+        globe,
       );
       if (defined(environmentState.skyAtmosphereCommand)) {
         this.updateDerivedCommands(environmentState.skyAtmosphereCommand);
@@ -3289,12 +3348,12 @@ Scene.prototype.updateEnvironment = function () {
   environmentState.isSunVisible = this.isVisible(
     environmentState.sunDrawCommand,
     cullingVolume,
-    occluder
+    occluder,
   );
   environmentState.isMoonVisible = this.isVisible(
     environmentState.moonCommand,
     cullingVolume,
-    occluder
+    occluder,
   );
 
   const envMaps = this.specularEnvironmentMaps;
@@ -3439,9 +3498,8 @@ function updateAndClearFramebuffers(scene, passState, clearColor) {
 
   // Update globe depth rendering based on the current context and clear the globe depth framebuffer.
   // Globe depth is copied for the pick pass to support picking batched geometries in GroundPrimitives.
-  const useGlobeDepthFramebuffer = (environmentState.useGlobeDepthFramebuffer = defined(
-    view.globeDepth
-  ));
+  const useGlobeDepthFramebuffer = (environmentState.useGlobeDepthFramebuffer =
+    defined(view.globeDepth));
   if (useGlobeDepthFramebuffer) {
     view.globeDepth.update(
       context,
@@ -3449,7 +3507,7 @@ function updateAndClearFramebuffers(scene, passState, clearColor) {
       view.viewport,
       scene.msaaSamples,
       scene._hdr,
-      environmentState.clearGlobeDepth
+      environmentState.clearGlobeDepth,
     );
     view.globeDepth.clear(context, passState, clearColor);
   }
@@ -3464,7 +3522,7 @@ function updateAndClearFramebuffers(scene, passState, clearColor) {
       passState,
       view.globeDepth.colorFramebufferManager,
       scene._hdr,
-      scene.msaaSamples
+      scene.msaaSamples,
     );
     oit.clear(context, passState, clearColor);
     environmentState.useOIT = oit.isSupported();
@@ -3484,7 +3542,7 @@ function updateAndClearFramebuffers(scene, passState, clearColor) {
       context,
       view.viewport,
       scene._hdr,
-      scene.msaaSamples
+      scene.msaaSamples,
     );
     view.sceneFramebuffer.clear(context, passState, clearColor);
 
@@ -3524,7 +3582,7 @@ function updateAndClearFramebuffers(scene, passState, clearColor) {
       scene._invertClassification.update(
         context,
         scene.msaaSamples,
-        view.globeDepth.colorFramebufferManager
+        view.globeDepth.colorFramebufferManager,
       );
       scene._invertClassification.clear(context, passState);
 
@@ -3534,7 +3592,7 @@ function updateAndClearFramebuffers(scene, passState, clearColor) {
         derivedCommands.oit = oit.createDerivedCommands(
           command,
           context,
-          derivedCommands.oit
+          derivedCommands.oit,
         );
       }
     } else {
@@ -3547,7 +3605,7 @@ function updateAndClearFramebuffers(scene, passState, clearColor) {
       scene._hdr,
       view.viewport,
       context,
-      passState
+      passState,
     );
   }
 }
@@ -3602,7 +3660,7 @@ Scene.prototype.resolveFramebuffers = function (passState) {
     const idTexture = idFramebuffer.getColorTexture(0);
     const depthTexture = defaultValue(
       globeFramebuffer,
-      sceneFramebuffer
+      sceneFramebuffer,
     ).getDepthStencilTexture();
     postProcess.execute(context, colorTexture, depthTexture, idTexture);
     postProcess.copy(context, defaultFramebuffer);
@@ -3712,7 +3770,7 @@ const updateHeightScratchCartographic = new Cartographic();
 Scene.prototype.updateHeight = function (
   cartographic,
   callback,
-  heightReference
+  heightReference,
 ) {
   //>>includeStart('debug', pragmas.debug);
   Check.typeOf.func("callback", callback);
@@ -3740,7 +3798,7 @@ Scene.prototype.updateHeight = function (
   if (!ignoreTerrain && defined(this.globe)) {
     terrainRemoveCallback = this.globe._surface.updateHeight(
       cartographic,
-      callbackWrapper
+      callbackWrapper,
     );
   }
 
@@ -3758,7 +3816,7 @@ Scene.prototype.updateHeight = function (
     const tilesetRemoveCallback = primitive.updateHeight(
       cartographic,
       callbackWrapper,
-      ellipsoid
+      ellipsoid,
     );
     tilesetRemoveCallbacks[primitive.id] = tilesetRemoveCallback;
   };
@@ -3772,10 +3830,10 @@ Scene.prototype.updateHeight = function (
   }
 
   const removeAddedListener = this.primitives.primitiveAdded.addEventListener(
-    createPrimitiveEventListener
+    createPrimitiveEventListener,
   );
-  const removeRemovedListener = this.primitives.primitiveRemoved.addEventListener(
-    (primitive) => {
+  const removeRemovedListener =
+    this.primitives.primitiveRemoved.addEventListener((primitive) => {
       if (primitive.isDestroyed() || !primitive.isCesium3DTileset) {
         return;
       }
@@ -3783,13 +3841,12 @@ Scene.prototype.updateHeight = function (
         tilesetRemoveCallbacks[primitive.id]();
       }
       delete tilesetRemoveCallbacks[primitive.id];
-    }
-  );
+    });
 
   const removeCallback = () => {
     terrainRemoveCallback = terrainRemoveCallback && terrainRemoveCallback();
     Object.values(tilesetRemoveCallbacks).forEach((tilesetRemoveCallback) =>
-      tilesetRemoveCallback()
+      tilesetRemoveCallback(),
     );
     tilesetRemoveCallbacks = {};
     removeAddedListener();
@@ -3854,7 +3911,7 @@ Scene.prototype.initializeFrame = function () {
         }
 
         this._globeHeight = updatedCartographic.height;
-      }
+      },
     );
   }
   this._cameraUnderground = isCameraUnderground(this);
@@ -3890,7 +3947,7 @@ function updateDebugShowFramesPerSecond(scene, renderedThisFrame) {
     scene._performanceDisplay =
       scene._performanceDisplay && scene._performanceDisplay.destroy();
     scene._performanceContainer.parentNode.removeChild(
-      scene._performanceContainer
+      scene._performanceContainer,
     );
   }
 }
@@ -4051,7 +4108,7 @@ Scene.prototype.render = function (time) {
     defined(this._lastRenderTime)
   ) {
     const difference = Math.abs(
-      JulianDate.secondsDifference(this._lastRenderTime, time)
+      JulianDate.secondsDifference(this._lastRenderTime, time),
     );
     shouldRender = shouldRender || difference > this.maximumRenderTimeChange;
   }
@@ -4065,7 +4122,7 @@ Scene.prototype.render = function (time) {
     const frameNumber = CesiumMath.incrementWrap(
       frameState.frameNumber,
       15000000.0,
-      1.0
+      1.0,
     );
     updateFrameNumber(this, frameNumber, time);
     frameState.newFrame = true;
@@ -4141,7 +4198,7 @@ Scene.prototype.requestRender = function () {
 Scene.prototype.clampLineWidth = function (width) {
   return Math.max(
     ContextLimits.minimumAliasedLineWidth,
-    Math.min(width, ContextLimits.maximumAliasedLineWidth)
+    Math.min(width, ContextLimits.maximumAliasedLineWidth),
   );
 };
 
@@ -4204,7 +4261,7 @@ Scene.prototype.pickVoxel = function (windowPosition, width, height) {
     this,
     windowPosition,
     width,
-    height
+    height,
   );
   // Look up the keyframeNode containing this picked cell
   const tileIndex = 255 * voxelCoordinate[0] + voxelCoordinate[1];
@@ -4220,7 +4277,7 @@ Scene.prototype.pickVoxel = function (windowPosition, width, height) {
     voxelPrimitive,
     tileIndex,
     sampleIndex,
-    keyframeNode
+    keyframeNode,
   );
 };
 
@@ -4243,12 +4300,12 @@ Scene.prototype.pickVoxel = function (windowPosition, width, height) {
  */
 Scene.prototype.pickPositionWorldCoordinates = function (
   windowPosition,
-  result
+  result,
 ) {
   return this._picking.pickPositionWorldCoordinates(
     this,
     windowPosition,
-    result
+    result,
   );
 };
 
@@ -4325,7 +4382,7 @@ function updateRequestRenderModeDeferCheckPass(scene) {
   // Check if any ignored requests are ready to go (to wake rendering up again)
   scene.primitives.updateForPass(
     scene._frameState,
-    requestRenderModeDeferCheckPassState
+    requestRenderModeDeferCheckPassState,
   );
 }
 
@@ -4377,14 +4434,14 @@ Scene.prototype.drillPickFromRay = function (
   ray,
   limit,
   objectsToExclude,
-  width
+  width,
 ) {
   return this._picking.drillPickFromRay(
     this,
     ray,
     limit,
     objectsToExclude,
-    width
+    width,
   );
 };
 
@@ -4404,13 +4461,13 @@ Scene.prototype.drillPickFromRay = function (
 Scene.prototype.pickFromRayMostDetailed = function (
   ray,
   objectsToExclude,
-  width
+  width,
 ) {
   return this._picking.pickFromRayMostDetailed(
     this,
     ray,
     objectsToExclude,
-    width
+    width,
   );
 };
 
@@ -4432,14 +4489,14 @@ Scene.prototype.drillPickFromRayMostDetailed = function (
   ray,
   limit,
   objectsToExclude,
-  width
+  width,
 ) {
   return this._picking.drillPickFromRayMostDetailed(
     this,
     ray,
     limit,
     objectsToExclude,
-    width
+    width,
   );
 };
 
@@ -4504,14 +4561,14 @@ Scene.prototype.clampToHeight = function (
   cartesian,
   objectsToExclude,
   width,
-  result
+  result,
 ) {
   return this._picking.clampToHeight(
     this,
     cartesian,
     objectsToExclude,
     width,
-    result
+    result,
   );
 };
 
@@ -4546,13 +4603,13 @@ Scene.prototype.clampToHeight = function (
 Scene.prototype.sampleHeightMostDetailed = function (
   positions,
   objectsToExclude,
-  width
+  width,
 ) {
   return this._picking.sampleHeightMostDetailed(
     this,
     positions,
     objectsToExclude,
-    width
+    width,
   );
 };
 
@@ -4586,13 +4643,13 @@ Scene.prototype.sampleHeightMostDetailed = function (
 Scene.prototype.clampToHeightMostDetailed = function (
   cartesians,
   objectsToExclude,
-  width
+  width,
 ) {
   return this._picking.clampToHeightMostDetailed(
     this,
     cartesians,
     objectsToExclude,
-    width
+    width,
   );
 };
 
@@ -4688,15 +4745,14 @@ function setTerrain(scene, terrain) {
   }
   // Otherwise, set a placeholder
   scene.globe.terrainProvider = undefined;
-  scene._removeTerrainProviderReadyListener = terrain.readyEvent.addEventListener(
-    (provider) => {
+  scene._removeTerrainProviderReadyListener =
+    terrain.readyEvent.addEventListener((provider) => {
       if (defined(scene) && defined(scene.globe)) {
         scene.globe.terrainProvider = provider;
       }
 
       scene._removeTerrainProviderReadyListener();
-    }
-  );
+    });
 }
 
 /**
@@ -4806,7 +4862,7 @@ Scene.prototype.destroy = function () {
     this._performanceDisplay =
       this._performanceDisplay && this._performanceDisplay.destroy();
     this._performanceContainer.parentNode.removeChild(
-      this._performanceContainer
+      this._performanceContainer,
     );
   }